big int

+++ backend/app/models.py
@@ -18,7 +18,7 @@ class FoodLog(Base):
     fat = Column(Float)
     carbs = Column(Float)
 
-    user_id = Column(Integer, ForeignKey("users.id"))
+    user_id = Column(BigInteger, ForeignKey("users.id"))
     user = relationship("User", back_populates="food_logs")
